Add certbot to provision ssl certs on creation.
This does not yet use the generated ssl certs. (That's pending.)

Certbot will wait for nginx to come up with the default config (that
contains the acme-challenge endpoint) and then request a certificate
with certonly in the init script.

Now we need a script in the nginx container to switch to the ssl config
and restart once the ssl certificate is obtained.